requirements.txt
Django
djangorestframework-simplejwt
markdown
django-filter
celery